pyproject.toml
[tool.poetry]
name = "koala-cli"
version = "0.1.0"
description = "Tool to update, install and healthcheck KoalaVim"
authors = ["Ofir Gal <ofir@gal.software>", "Michael Shustin <michaelshustin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
typer = {extras = ["all"], version = "^0.9.0"}
gitpython = "^3.1.40"
typing-extensions = "^4.9.0"
patool = "^2.1.0"

[tool.poetry.scripts]
koala = "koala_cli.__main__:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
skip-string-normalization = true
target-version = ['py38']
